pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "buch-ai-server"
version = "0.1.0"
description = "A FastAPI application."
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
requires-python = ">= 3.12"
dependencies = [
    "aiohttp==3.11.12",
    "bcrypt==4.0.1",
    "faiss-cpu>=1.11.0",
    "fastapi>=0.95.0",
    "fastapi[standard]>=0.95.0",
    "firebase-admin>=6.9.0",
    "google-cloud-aiplatform>=1.86.0",
    "google-cloud-bigquery>=3.11.0",
    "google-cloud-firestore>=2.21.0",
    "google-cloud-run>=0.10.17",
    "google-cloud-texttospeech>=2.26.0",
    "google-genai>=1.19.0",
    "httpx>=0.28.1",
    "huggingface-hub>=0.29.2",
    "moviepy==2.2.1",
    "numpy>=2.2.4",
    "passlib==1.7.4",
    "pillow==11.2.1",
    "pydantic>=2.10.6",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "setuptools>=80.9.0",
    "stripe>=7.0.0",
    "uv-sort==0.6.0",
    "uvicorn>=0.22.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pytest>=7.2.0",
    "ruff>=0.11.1",
]
mcp = [
    "mcp[cli]>=1.9.2",
]
